+++ src/nodes/logic/index.ts
@@ -1,2 +1,3 @@
+export * from "./extractCoordinates";
 export * from "./massLoop";
 export * from "./moduleEnabled";