+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.9.7
+- (Nodes) Added an Extract Coordinates node to capture a target's current position for later use.
 - (Nodes) Added a text input to Location I/O for use with "Named Location" node.
 - (Nodes) Added a text output to "Named Location" node.